refactor: separate sessions spec tests from driver specific tests
