+++ .evergreen/run-serverless-tests.sh
@@ -16,7 +16,7 @@ npx mocha --file test/tools/runner/index.js \
   test/integration/crud/crud.prose.test.js \
   test/integration/retryable-reads/retryable_reads.spec.test.js \
   test/integration/retryable-writes/retryable_writes.spec.test.js \
-  test/functional/sessions.test.js \
+  test/integration/sessions/sessions.spec.test.js \
   test/functional/transactions.test.js \
   test/integration/versioned-api/versioned_api.spec.test.js \
   test/integration/load-balancers/load_balancers.spec.test.js
